+++ test/multi-select/multi-select-item/multi-select-item.component.spec.js
@@ -6,14 +6,13 @@ import React from 'react';
 import { expect } from 'chai';
 import { mount } from 'enzyme';
 
-import MultiSelectItem from '../../../src/multi-select/multi-select-item/multi-select-item.component.jsx';
-
+import MultiSelectItem from '../../../src/multi-select/multi-select-item/multi-select-item.component';
 
 describe('MultiSelectItem component', function describe() {
   it('should render correctly', function it() {
     const props = {
       item: {
-        value: 1,
+        value: 'Item_1',
         label: 'Item 1',
       },
       isChecked: true,
@@ -23,7 +22,7 @@ describe('MultiSelectItem component', function describe() {
       <MultiSelectItem {...props} />,
     );
 
-    expect(wrapper.find('.oc-multi-select-item-label').at(0).text()).to.eql('Item 1');
-    expect(wrapper.find('#1').props().checked).to.be.true;
+    expect(wrapper.find('.oc-multi-select-item-checkbox span').at(0).text()).to.eql('Item 1');
+    expect(wrapper.find('#Item_1').at(0).props().checked).to.be.true;
   });
 });